+++ android_vendor_itinerary/app/src/main/java/com/bhargav/tool/ModelClass/FacturasXCobrarTemp.java
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName;
 
 public class FacturasXCobrarTemp {
 
-    private int valor = 0;
+    private double valor = 0.0;
     private int codigoEmpresa = 1;
     private int secuenciaComprobante;
     private int isChecked = 0;
@@ -34,11 +34,11 @@ public class FacturasXCobrarTemp {
         this.codigoEmpresa = codigoEmpresa;
     }
 
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 }
